Translate internal role to user-facing role; Add grid search capabilities; Add node status to grid page; Implement manager status framework with push-to-browser capabilities, websocket disconnection detection, title bar indications

+++ module/options.go
@@ -38,7 +38,7 @@ func GetInt(options map[string]interface{}, key string) (int, error) {
   var err error
   var value int
   if gen, ok := options[key]; ok {
-    value = gen.(int)
+    value = int(gen.(float64))
   } else {
     err = errors.New("Required option is missing: " + key + " (int)")
   }
@@ -48,7 +48,7 @@ func GetInt(options map[string]interface{}, key string) (int, error) {
 func GetIntDefault(options map[string]interface{}, key string, dflt int) int {
   var value int
   if gen, ok := options[key]; ok {
-    value = gen.(int)
+    value = int(gen.(float64))
   } else {
     value = dflt
   }
